server_backend/requirements.txt
Flask
mysql-connector
pymysql
requests
